requirements.txt
google-cloud-storage==2.18.2 # Google Cloud Storage client
google-cloud-bigquery==3.26.0 # Google BigQuery client
google-cloud-aiplatform==1.71.1 # Google Vertex AI client

pyyaml==6.0.2
pytest==8.3.4

scikit-learn==1.6.0

tensorflow==2.18.0

fastapi
uvicorn